Former-commit-id: 9d5a0ed8c13ff328a31b7c20179433f3a082751a
Former-commit-id: 93940ca132c24e7fac19b979438dfa83b55a1260

+++ src/fswAlgorithms/attGuidance/locationPointing/locationPointing.h
@@ -41,7 +41,7 @@ typedef struct {
 
     /* private variables */
     double sigma_BR_old[3];     /*!< Older sigma_BR value, stored for finite diff*/
-    double time_old;            /*!< prior time value */
+    uint64_t time_old;          /*!< [ns] prior time value */
     double init;                /*!< moudle initialization counter */
     double eHat180_B[3];        /*!< -- Eigen axis to use if commanded axis is 180 from pHat */
